Move all into same SEO Component 🌊
