+++ src/components/Seo.jsx
@@ -1,0 +1,111 @@
+/**
+ * SEO component that queries for data with
+ *  Gatsby's useStaticQuery React hook
+ *
+ * See: https://www.gatsbyjs.com/docs/use-static-query/
+ */
+
+import * as React from "react";
+import PropTypes from "prop-types";
+import { Helmet } from "react-helmet";
+import { useLocation } from "@reach/router";
+import { useStaticQuery, graphql } from "gatsby";
+
+import logo from "../images/icon.png";
+
+const Seo = ({ description, lang, meta, title }) => {
+    const { site } = useStaticQuery(
+        graphql`
+            query {
+                site {
+                    siteMetadata {
+                        title
+                        description
+                        siteUrl
+                    }
+                }
+            }
+        `
+    );
+
+    const { pathname } = useLocation();
+    const metaDescription = description || site.siteMetadata.description;
+    const defaultTitle = site.siteMetadata?.title;
+    const url = `${site.siteMetadata.siteUrl}${pathname}`;
+
+    return (
+        <Helmet
+            htmlAttributes={{
+                lang,
+            }}
+            title={title}
+            titleTemplate={defaultTitle ? `%s | ${defaultTitle}` : null}
+            meta={[
+                {
+                    property: `title`,
+                    content: title,
+                },
+                {
+                    name: `description`,
+                    content: metaDescription,
+                },
+                {
+                    property: `og:url`,
+                    content: url,
+                },
+                {
+                    property: `og:title`,
+                    content: title,
+                },
+                {
+                    property: `og:description`,
+                    content: metaDescription,
+                },
+                {
+                    property: `og:type`,
+                    content: `website`,
+                },
+                {
+                    property: `og:image`,
+                    content: logo,
+                },
+                {
+                    property: `twitter:url`,
+                    content: url,
+                },
+                {
+                    name: `twitter:card`,
+                    content: `summary_large_image`,
+                },
+                {
+                    name: `twitter:title`,
+                    content: title,
+                },
+
+                {
+                    name: `twitter:description`,
+                    content: metaDescription,
+                },
+                {
+                    property: `twitter:image`,
+                    content: logo,
+                },
+            ].concat(meta)}
+        />
+    );
+};
+
+Seo.defaultProps = {
+    lang: `en`,
+    meta: [],
+    description: ``,
+};
+
+Seo.propTypes = {
+    description: PropTypes.string,
+    lang: PropTypes.string,
+    meta: PropTypes.arrayOf(PropTypes.object),
+    title: PropTypes.string.isRequired,
+};
+
+export default Seo;
