+++ src/pages/blog.js
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { Link, graphql } from "gatsby";
 
-import Layout from "../components/layout";
+import Layout from "../components/Layout";
 import Seo from "../components/Seo";
 
 const BlogIndex = ({ data, location }) => {
