+++ src/pages/index.tsx
@@ -1,33 +1,28 @@
 import * as React from "react";
-import { Link, graphql } from "gatsby";
-
+import { useStaticQuery, graphql } from "gatsby";
 
 import FadeIn from "../components/fadeIn";
-import Head from "../components/Head";
-import Footer from "../components/Footer";
-import Header from "../components/Header";
-import Layout from "../components/layout";
+import Layout from "../components/Layout";
+import Seo from "../components/Seo";
 
 // markup
-const IndexPage = ({  location }) => {
-    const siteTitle =  `Title`;
+const IndexPage = ({ location }) => {
+    const siteTitle = "Home";
 
     return (
-        <>
-            <Head />
-            <Layout location={location} title={siteTitle}>
-                <FadeIn delay={100} duration={1000}>
-                    <h1>Hunter Software Development</h1>
-                    <h4>
-                        A Boutique Web Development Agency in Greenville, South Carolina
-                        <span role="img" aria-label="pin">
-                            📍
-                        </span>
-                    </h4>
-                    <p>Coming soon...</p>
-                </FadeIn>
-            </Layout>
-        </>
+        <Layout location={location} title={siteTitle}>
+            <Seo title={siteTitle} />
+            <FadeIn delay={100} duration={1000}>
+                <h1>Hunter Software Development</h1>
+                <h4>
+                    A Boutique Web Development Agency in Greenville, South Carolina
+                    <span role="img" aria-label="pin">
+                        📍
+                    </span>
+                </h4>
+                <p>Coming soon...</p>
+            </FadeIn>
+        </Layout>
     );
 };
